// src/Entities/Profile/model/types/profile.ts
import { Country } from 'Entities/Country';
import { Currency } from 'Entities/Currency';

export enum ValidateProfileErrors {
   INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
   INCORRECT_AGE = 'INCORRECT_AGE',
   INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
   NO_DATA = 'NO_DATA',
   SERVER_ERROR = 'SERVER_ERROR',
}

export interface Profile {
   first?: string;
   lastname?: string;
   age?: number;
   currency?: Currency;
   country?: Country;
   city?: string;
   username?: string;
   avatar?: string;
}

export interface ProfileSchema {
   data?: Profile;
   form?: Profile;
   isLoading: boolean;
   error?: string;
   readonly: boolean;
   validateErrors?: ValidateProfileErrors[];
}